feat: extracted out new widget from video feed
